LMS1-587: validate term deletion (#489)

+++ src/main/java/uk/co/nstauthority/licensingmanagementservice/licence/schedule/licencescheduleterm/LicenceScheduleTermDeletionController.java
@@ -1,11 +1,13 @@
 package uk.co.nstauthority.licensingmanagementservice.licence.schedule.licencescheduleterm;
 
 import java.util.UUID;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import uk.co.nstauthority.licensingmanagementservice.authorisation.HasRolesInTeamType;
@@ -14,6 +16,7 @@ import uk.co.nstauthority.licensingmanagementservice.authorisation.rules.licence
 import uk.co.nstauthority.licensingmanagementservice.fds.notificationbanner.NotificationBanner;
 import uk.co.nstauthority.licensingmanagementservice.licence.LicenceService;
 import uk.co.nstauthority.licensingmanagementservice.licence.schedule.calculation.LicenceScheduleCalculationService;
+import uk.co.nstauthority.licensingmanagementservice.licence.schedule.common.ScheduleRelativeDateValidationService;
 import uk.co.nstauthority.licensingmanagementservice.licence.schedule.eventcomments.EventComment;
 import uk.co.nstauthority.licensingmanagementservice.licence.schedule.eventcomments.EventCommentService;
 import uk.co.nstauthority.licensingmanagementservice.licence.schedule.licencescheduledetail.LicenceScheduleDetailStatus;
@@ -34,17 +37,20 @@ public class LicenceScheduleTermDeletionController {
   private final LicenceScheduleCalculationService licenceScheduleCalculationService;
   private final EventCommentService eventCommentService;
   private final LicenceService licenceService;
+  private final ScheduleRelativeDateValidationService scheduleRelativeDateValidationService;
 
   public LicenceScheduleTermDeletionController(
       LicenceScheduleTermService licenceScheduleTermService,
       LicenceScheduleCalculationService licenceScheduleCalculationService,
       EventCommentService eventCommentService,
-      LicenceService licenceService
+      LicenceService licenceService,
+      ScheduleRelativeDateValidationService scheduleRelativeDateValidationService
   ) {
     this.licenceScheduleTermService = licenceScheduleTermService;
     this.licenceScheduleCalculationService = licenceScheduleCalculationService;
     this.eventCommentService = eventCommentService;
     this.licenceService = licenceService;
+    this.scheduleRelativeDateValidationService = scheduleRelativeDateValidationService;
   }
 
   @GetMapping
@@ -58,6 +64,13 @@ public class LicenceScheduleTermDeletionController {
             .orElse("")
         : "";
 
+    var canDeleteErrorMessage = "";
+    if (!licenceScheduleTermService.canDeleteTerm(term)) {
+      canDeleteErrorMessage = "This cannot be deleted as it is referenced by other schedule events.";
+    } else if (!scheduleRelativeDateValidationService.isTermRemovalValid(term)) {
+      canDeleteErrorMessage = "This cannot be deleted as it would cause events to end after the final term.";
+    }
+
     return new ModelAndView("lms/licence/schedule/deleteScheduleTerm")
         .addObject("pageTitle", PAGE_TITLE.formatted(term.getTermType().getDisplayName()))
         .addObject("licenceScheduleTermSummaryView", LicenceScheduleTermSummaryView.fromTerm(term))
@@ -65,7 +78,7 @@ public class LicenceScheduleTermDeletionController {
         .addObject("cancelUrl", term.getLicenceScheduleDetail().getScheduleTimelineRouteUrl())
         .addObject("pageCaption",
             licenceService.getLicencePageCaption(term.getLicenceScheduleDetail().getLicenceSchedule().getLicence()))
-        .addObject("canDelete", licenceScheduleTermService.canDeleteTerm(term));
+        .addObject("canDeleteErrorMessage", canDeleteErrorMessage);
   }
 
   @PostMapping
@@ -75,6 +88,21 @@ public class LicenceScheduleTermDeletionController {
   ) {
     var term = licenceScheduleTermService.getTermByIdOrThrow(licenceScheduleTermId);
     var licenceScheduleDetail = term.getLicenceScheduleDetail();
+
+    if (!licenceScheduleTermService.canDeleteTerm(term)) {
+      throw new ResponseStatusException(
+          HttpStatus.FORBIDDEN,
+          "Cannot delete term id: %s as it is referenced by other schedule events".formatted(term.getId())
+      );
+    }
+
+    if (!scheduleRelativeDateValidationService.isTermRemovalValid(term)) {
+      throw new ResponseStatusException(
+          HttpStatus.FORBIDDEN,
+          "Cannot delete term id: %s as it would cause events to occur after the end of the final term".formatted(term.getId())
+      );
+    }
+
     licenceScheduleTermService.deleteTerm(term);
 
     licenceScheduleCalculationService.calculateAndSaveLicenceScheduleDates(licenceScheduleDetail);
